ENH: add new common base class for psana bases
